Enhance i/o tensor to buffer conversion in HLOToLinalgOnBuffers pass.
The current handling of converting i/o tensors to buffers works well
when there is a single operation within the dispatch function. As the
codegen evolves to handle tiling and fusion at buffer level, this
would be fragile. This makes the handling better, but more
enhancements might be needed here.
Also various cleanup within the pass.

PiperOrigin-RevId: 311377587